fix: run checks after committing first time round
