+++ .github/workflows/manual-run-force.yml
@@ -56,3 +56,15 @@ jobs:
           commit_author: "fontsource-bot <83556432+fontsource-bot@users.noreply.github.com>"
           commit_message: "chore(build): update packages [Force Rebuild]"
         continue-on-error: true
+
+      - name: Check font files # Detects if all binaries are downloaded successfully and in the right place
+        run: yarn util:run-check
+
+      - name: If error, commit again before publishing
+        uses: stefanzweifel/git-auto-commit-action@v4
+        with:
+          commit_user_name: "fontsource-bot"
+          commit_user_email: "83556432+fontsource-bot@users.noreply.github.com"
+          commit_author: "fontsource-bot <83556432+fontsource-bot@users.noreply.github.com>"
+          commit_message: "chore(build): resolve file check errors [Force Rebuild]"
+        continue-on-error: true
