+++ .github/workflows/manual-run.yml
@@ -57,6 +57,18 @@ jobs:
           commit_message: "chore(build): update packages [Manual]"
         continue-on-error: true
 
+      - name: Check font files # Detects if all binaries are downloaded successfully and in the right place
+        run: yarn util:run-check
+
+      - name: If error, commit again before publishing
+        uses: stefanzweifel/git-auto-commit-action@v4
+        with:
+          commit_user_name: "fontsource-bot"
+          commit_user_email: "83556432+fontsource-bot@users.noreply.github.com"
+          commit_author: "fontsource-bot <83556432+fontsource-bot@users.noreply.github.com>"
+          commit_message: "chore(build): resolve file check errors [Manual]"
+        continue-on-error: true
+
       - name: Configure CI Git
         run: |
           git config --global user.email "83556432+fontsource-bot@users.noreply.github.com"
