+++ package.json
@@ -30,6 +30,7 @@
     "readme-update:google": "ts-node scripts/google/google-readme-updater.ts",
     "util:algolia": "ts-node scripts/utils/algolia-index.ts",
     "util:fontlist": "ts-node scripts/utils/fontlist-generator.ts",
+    "util:run-check": "ts-node scripts/google/run-checks.ts",
     "util:generic-renamer": "ts-node scripts/generic/renamer.ts",
     "util:package-json-rebuild": "ts-node scripts/utils/package-json-rebuild.ts",
     "util:subsets": "ts-node scripts/utils/subset-list.ts",
