+++ scripts/google/new-font-check.ts
@@ -1,10 +1,17 @@
 import fs from "node:fs";
+
 import _ from "lodash";
 import path from "node:path";
 import { getDirectories } from "scripts/utils/utils";
 
-// Google may sometimes push a new font that already exists in the generic folder
-// This checks if there are any duplicates between the two font folders and purges the duplicate from generic
+/**
+ * Google may sometimes push a new font that already exists in the generic folder
+ * This checks if there are any duplicates between the two font folders and purges the duplicate from generic
+ */
+
+/**
+ * Gets all directories in all font folders
+ */
 const directories = [
   ...getDirectories("google"),
   ...getDirectories("league"),
@@ -12,7 +19,11 @@ const directories = [
   ...getDirectories("other"),
 ];
 
-// Finds duplicates between the directories
+/**
+ * Find all the duplicate values in the given array
+ * @param dirs array of directory names
+ * @returns the duplicate values
+ */
 const findDuplicates = (dirs: string[]) =>
   _.filter(dirs, (value, index, iteratee) =>
     _.includes(iteratee, value, index + 1)
@@ -20,7 +31,11 @@ const findDuplicates = (dirs: string[]) =>
 
 const duplicates = findDuplicates(directories);
 
-// Delete all duplicate dirs
+/**
+ * Delete all the duplicate directories from the font folder
+ * @param duplicateDirs array of directory names
+ * @returns
+ */
 const deleteDuplicates = (duplicateDirs: string[]): string[] => {
   duplicateDirs.forEach(dir => {
     try {
